Add new main.css and main.js files
Included a new CSS file, 'main.css', with styles for various elements and components of the website. Also added a new JavaScript file, 'main.js', for enhancing the navigation functionality of the site. These new files aim to improve the overall look and feel and enhance user experience on the website.