Implement human-friendly date formatting
Added a function to display dates in a more human-friendly format across various templates. The newly created "humanDate" function converts a timestamp into a more readable format and is now used to display 'created' and 'expires' dates. In addition, updated the Go version in go.mod and inserted the current year in the footer of the base HTML template.